qr code scanner code global to local and some optimmisation
