+++ lib/app/interop/issuer/models/organization_info.dart
