+++ lib/qr_code/qr_code_scan/qr_code_scan.dart
@@ -1,2 +1,2 @@
-export 'view/qr_code_scan.dart';
-export 'cubit/qr_code_cubit.dart';
+export 'view/qr_code_scan_page.dart';
+export 'cubit/qr_code_scan_cubit.dart';
