--- lib/qr_code/qr_code_scan/view/qr_code_scan.dart
@@ -1,96 +1,0 @@
-import 'dart:io';
-
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:qr_code_scanner/qr_code_scanner.dart';
-import 'package:talao/app/shared/widget/back_leading_button.dart';
-import 'package:talao/app/shared/widget/base/page.dart';
-import 'package:talao/qr_code/qr_code_scan/cubit/qr_code_cubit.dart';
-import 'package:talao/qr_code/qr_code_scan/cubit/qr_code_state.dart';
-
-class QrCodeScanPage extends StatefulWidget {
-  static Route route() => MaterialPageRoute(
-        builder: (context) => QrCodeScanPage(),
-        settings: RouteSettings(name: '/qrCodeScanPage'),
-      );
-
-  @override
-  _QrCodeScanPageState createState() => _QrCodeScanPageState();
-}
-
-class _QrCodeScanPageState extends State<QrCodeScanPage> {
-  final qrKey = GlobalKey(debugLabel: 'QR');
-  late QRViewController qrController;
-
-  late bool flash;
-  bool promptActive = false;
-
-  @override
-  void initState() {
-    super.initState();
-    flash = false;
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
-    qrController.dispose();
-  }
-
-  @override
-  void reassemble() {
-    super.reassemble();
-    if (Platform.isAndroid) {
-      qrController.pauseCamera();
-    } else if (Platform.isIOS) {
-      qrController.resumeCamera();
-    }
-  }
-
-  void onQRViewCreated(QRViewController controller) {
-    qrController = controller;
-
-    qrController.scannedDataStream.listen((scanData) {
-      qrController.pauseCamera();
-      if (scanData.code is String && !promptActive) {
-        setState(() {
-          promptActive = true;
-        });
-        context.read<QRCodeCubit>().host(scanData.code);
-      }
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final localizations = AppLocalizations.of(context)!;
-    return BlocListener<QRCodeCubit, QRCodeState>(
-      listener: (context, state) {
-        if (state is QRCodeStateUnknown) {
-          qrController.resumeCamera();
-
-          ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-            content: Text(localizations.scanUnsupportedMessage),
-          ));
-        }
-      },
-      child: BasePage(
-        padding: EdgeInsets.zero,
-        title: localizations.scanTitle,
-        scrollView: false,
-        extendBelow: true,
-        titleLeading: BackLeadingButton(),
-        body: SafeArea(
-          child: QRView(
-            key: qrKey,
-            overlay: QrScannerOverlayShape(
-              borderColor: Colors.white70,
-            ),
-            onQRViewCreated: onQRViewCreated,
-          ),
-        ),
-      ),
-    );
-  }
-}
